user moved to people
