--- trojsten/admin.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from __future__ import unicode_literals
-
-from django.contrib import admin
-from models import User
-
-admin.site.register(User)
